+++ src/main/java/net/semanticmetadata/lire/solr/FeatureRegistry.java
@@ -1,9 +1,12 @@
 package net.semanticmetadata.lire.solr;
 
+import net.semanticmetadata.lire.imageanalysis.features.GenericDoubleLireFeature;
 import net.semanticmetadata.lire.imageanalysis.features.GlobalFeature;
 import net.semanticmetadata.lire.imageanalysis.features.global.*;
 import net.semanticmetadata.lire.imageanalysis.features.global.joint.JointHistogram;
 import net.semanticmetadata.lire.imageanalysis.features.global.spatialpyramid.SPCEDD;
+import net.semanticmetadata.lire.solr.features.DoubleFeatureCosineDistance;
+import net.semanticmetadata.lire.solr.features.ShortFeatureCosineDistance;
 
 import java.util.HashMap;
 import java.util.Iterator;
@@ -50,6 +53,12 @@ public class FeatureRegistry {
         codeToClass.put("jh", JointHistogram.class);
         codeToClass.put("sc", ScalableColor.class);
         codeToClass.put("pc", SPCEDD.class);
+        // GenericFeatures filled with whatever one prefers.
+        codeToClass.put("df", DoubleFeatureCosineDistance.class);
+//        codeToClass.put("df", GenericGlobalDoubleFeature.class);
+        codeToClass.put("if", GenericGlobalIntFeature.class);
+        codeToClass.put("sf", ShortFeatureCosineDistance.class);
+//        codeToClass.put("sf", GenericGlobalShortFeature.class);
 
         // local feature based histograms.
         // codeToClass.put("sim_ce", GenericByteLireFeature.class); // SIMPLE CEDD ... just to give a hint how it might look like.