+++ src/main/java/net/semanticmetadata/lire/solr/indexing/ParallelSolrIndexer.java
@@ -61,8 +61,8 @@ import java.util.concurrent.LinkedBlockingQueue;
 
 /**
  * This indexing application allows for parallel extraction of global features from multiple image files for
- * use with the LIRE Solr plugin. It basically takes a list of images (ie. created by something like
- * "dir /s /b &gt; list.txt" or "ls [some parameters] &gt; list.txt".
+ * use with the LIRE Solr plugin. It basically takes a list of images, i.e. created by something like
+ * "dir /s /b &gt; list.txt" or "find /path/to/images -name "*.jpg" &gt; list.txt".
  *
  * use it like:
  * <pre>$&gt; java -jar lire-request-handler.jar -i &lt;infile&gt; [-o &lt;outfile&gt;] [-n &lt;threads&gt;] [-m &lt;max_side_length&gt;] [-f]</pre>
@@ -128,7 +128,7 @@ public class ParallelSolrIndexer implements Runnable {
     }
 
     public static void main(String[] args) throws IOException {
-        BitSampling.readHashFunctions();
+        HashingMetricSpacesManager.init();
         ParallelSolrIndexer e = new ParallelSolrIndexer();
 
         // parse programs args ...
@@ -444,7 +444,7 @@ public class ParallelSolrIndexer implements Runnable {
                 while ((file = br.readLine()) != null) {
                     next = new File(file);
                     try {
-                        // reading from harddrive to buffer to reduce the load on the HDD and move decoding to the
+                        // reading from hard drive to buffer to reduce the load on the HDD and move decoding to the
                         // consumers using java.nio
                         int fileSize = (int) next.length();
                         byte[] buffer = new byte[fileSize];
@@ -546,6 +546,9 @@ public class ParallelSolrIndexer implements Runnable {
                         }
                         // --------< creating doc >-------------------------
                         sb.append("<doc>");
+                        sb.append("<field name=\"localimagefile\">");
+                        sb.append(tmp.getFileName());
+                        sb.append("</field>");
                         sb.append("<field name=\"id\">");
                         if (idp == null)
                             sb.append(tmp.getFileName());